Create (NEW)AndroidAdapter.java
There could be a problem with the old files.  The imbedded calendar made a lot of stuff easy, but it also made adding events ridiculously hard, so I tried to redo the base calendar.  Any file with (NEW) is a complete redo.  I'll keep the old build on there just in case.