Create activity_week_view.xml
Made the days buttons that can lead to listView for that day, but that has to be set up.  The Week Of tab works like the month name on the original calendar, so the same setup to set the week start and end.  The prev and next buttons should work the same as well.